Create project for the RestTemplate

+++ project3/RestAPI/src/main/java/com/udemy/springcourse/controllers/MeasurementsController.java
@@ -42,7 +42,7 @@ public class MeasurementsController {
     }
 
     @GetMapping("/rainyDaysCount")
-    public int showRainyDays() {
+    public Long showRainyDays() {
         return measurementsService.countRainyDays();
     }
